+++ Changelog.md
@@ -6,6 +6,7 @@ WhateverGreen Changelog
 - Added `igfxfcms=1` boot argument (and `complete-modeset` property) on Skylake and Apple
 - Improved performance with Lilu 1.4.3 APIs
 - Added `-igfxfbdbg` boot argument to debug IGPU framebuffer (debug builds only)
+- Added `igfxagdc=0` boot argument and `disable-agdc` IGPU property to disable AGDC
 
 #### v1.3.7
 - Improved the maximum link rate fix: Now correct the value read from extended DPCD as well. (by @0xFireWolf)